+++ features/Context/Page/Base/Base.php
@@ -2,6 +2,7 @@
 
 namespace Context\Page\Base;
 
+use Behat\Mink\Element\NodeElement;
 use SensioLabs\Behat\PageObjectExtension\PageObject\Page;
 use Behat\Mink\Exception\UnsupportedDriverActionException;
 use Behat\Mink\Exception\ElementNotFoundException;
@@ -15,14 +16,14 @@ use Behat\Mink\Exception\ElementNotFoundException;
  */
 class Base extends Page
 {
-    protected $elements = array(
-        'Dialog'         => array('css' => 'div.modal'),
-        'Title'          => array('css' => '.navbar-title'),
-        'HeadTitle'      => array('css' => 'title'),
-        'Flash messages' => array('css' => '.flash-messages-holder'),
-        'Navigation Bar' => array('css' => 'header#oroplatform-header'),
-        'Container'      => array('css' => '#container'),
-    );
+    protected $elements = [
+        'Dialog'         => ['css' => 'div.modal'],
+        'Title'          => ['css' => '.navbar-title'],
+        'HeadTitle'      => ['css' => 'title'],
+        'Flash messages' => ['css' => '.flash-messages-holder'],
+        'Navigation Bar' => ['css' => 'header#oroplatform-header'],
+        'Container'      => ['css' => '#container'],
+    ];
 
     /**
      * {@inheritdoc}
@@ -83,7 +84,7 @@ class Base extends Page
      *
      * @return string
      */
-    public function getUrl(array $options = array())
+    public function getUrl(array $options = [])
     {
         $url = $this->getPath();
 
@@ -96,6 +97,9 @@ class Base extends Page
 
     /**
      * Get page title
+     *
+     * @throws \Exception
+     *
      * @return string
      */
     public function getTitle()
@@ -132,10 +136,10 @@ class Base extends Page
         if (!$button) {
             $button =  $this->find(
                 'named',
-                array(
+                [
                     'link',
                     $this->getSession()->getSelectorsHandler()->xpathLiteral($locator)
-                )
+                ]
             );
         }