+++ src/Pim/Bundle/DataGridBundle/Resources/public/js/datagrid/action-launcher.js
@@ -50,16 +50,16 @@ function($, _, Backbone) {
                 '<%= className ? " " + className : "" %>' +
                 '<%= !enabled ? " disabled" : "" %>' +
                 '"' +
+                ' data-toggle="tooltip" data-placement="top" title="<%= label %>"' +
                 ' <%= attributesTemplate({attributes: attributes}) %>' +
-                ' title="<%= label %>"' +
             '>' +
                 '<% if (icon) { %>' +
                     '<i class="icon-<%= icon %> hide-text"><%= label %></i>' +
                 '<% } else { %>' +
                     '<% if (iconClassName) { %>' +
-                        '<i class="<%= iconClassName %>"></i>' +
+                        '<i class="<%= iconClassName %>" ' +
+                        '></i>' +
                     '<% } %>' +
-                    ' <%= label %>' +
                 '<% } %>' +
             '</<%= tagName %>>'
         ),